refactor, api change: cpoisson_pmf to cpoisson_pdf
