+++ src/numba_stats/expon.py
@@ -1,0 +1,36 @@
+import numba as nb
+import numpy as np
+from ._special import expm1, log1p
+
+_signatures = [
+    nb.float32(nb.float32, nb.float32, nb.float32),
+    nb.float64(nb.float64, nb.float64, nb.float64),
+]
+
+
+@nb.vectorize(_signatures)
+def pdf(x, mu, sigma):
+    """
+    Return probability density of exponential distribution.
+    """
+    z = (x - mu) / sigma
+    return np.exp(-z) / sigma
+
+
+@nb.vectorize(_signatures)
+def cdf(x, mu, sigma):
+    """
+    Evaluate cumulative distribution function of exponential distribution.
+    """
+    z = (x - mu) / sigma
+    return -expm1(-z)
+
+
+@nb.vectorize(_signatures)
+def ppf(p, mu, sigma):
+    """
+    Return quantile of exponential distribution for given probability.
+    """
+    z = -log1p(-p)
+    x = z * sigma + mu
+    return x
